Removed entity_list drawing from graphics class
Removed the set_entity_list and draw() commands from the Graphics class,
and moved them into the EntityList class itself, using a reference to
the Graphics object to draw them instead.

Added a Makefile which runs the Makefile in it's parent directory. This
is used to build the project from inside the Src folder.

Removed line from previous code