Implimented basic Map
Implimented a basic Map, which now parses map.bmp in the data folder, and
creates tiles at locations related to the tiles/colors defined in tileset1.json.

Also reorganized Spritesheet so that Tileset and Animated Spritesheet both
are derived from it, therefore moving all animation related information
out of the class and into it's own.